плюсом убрал все инт64 на инт

+++ internal/domain/sectionItems/entity/sectionItems.go
@@ -1,8 +1,8 @@
 package entity
 
 type SectionItem struct {
-	SectionID int64  `json:"sectionID"`
+	SectionID int    `json:"sectionID"`
 	IsTest    bool   `json:"isTest"`
 	Title     string `json:"title"`
-	ItemID    int64  `json:"itemID"`
+	ItemID    int    `json:"itemID"`
 }
