+++ internal/domain/sections/entity/sections.go
@@ -1,6 +1,6 @@
 package entity
 
 type Sections struct {
-	ID    int64  `json:"ID"`
+	ID    int    `json:"ID"`
 	Title string `json:"title"`
 }
